test(naa): keep flow tests in dedicated specs

Copilot-Session: 80a81e22-5306-426c-b5e7-4ae919f1838d
